🩹(e2e) force English language before left-panel tests
