+++ src/frontend/apps/e2e/__tests__/app-conversations/projects.spec.ts
@@ -1,18 +1,8 @@
-import { Page, expect, test } from '@playwright/test';
+import { expect, test } from '@playwright/test';
+
+import { createProject } from '../helpers';
 
 import { randomName } from './common';
-
-const createProject = async (page: Page, projectName: string) => {
-  await page.getByRole('button', { name: 'New project' }).click();
-  const createModal = page.getByRole('dialog', {
-    name: 'Content modal to create a project',
-  });
-  await createModal
-    .getByRole('textbox', { name: 'Project name' })
-    .fill(projectName);
-  await createModal.getByRole('button', { name: 'Create project' }).click();
-  await expect(page.getByText('The project has been created.')).toBeVisible();
-};
 
 test.describe('Projects', () => {
   test.beforeEach(async ({ page }) => {
