+++ src/frontend/apps/e2e/package.json
@@ -4,6 +4,8 @@
   "private": true,
   "scripts": {
     "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "prettier": "prettier --write .",
     "install-playwright": "playwright install --with-deps",
     "test": "playwright test",
     "test:ui": "yarn test --ui",
